fix(audit): /code-audit pr 546 re-invocation 3 — partition reviewer findings
Pass 1 (single-pass invocation) reviewers found 2 HIGH (file-size cap
violations), ~17 MAJOR/CRITICAL correctness items, ~33 MEDIUM, ~50 LOW.
Applied load-bearing fixes; documented descope for remaining polish.

HIGH file-size cap closures:
- main.cpp 853→490 lines via cli_env / cli_io / cli_wait / cli_target TU
  split (each helper under 130 lines). Helpers in PhosphorPipeWireCli
  namespace to avoid collision with library public surface.
- test_palettestore.cpp 841→388 lines via sibling TU split + shared
  helpers header (test_palettestore_files.cpp + _hotreload.cpp +
  _helpers.h). Mirrors phosphor-registry tests/ pattern. seedTempJsonFile
  / writeJsonFile / defaultWrappedPayload helpers consolidate the
  per-test temp-JSON setup.

P3 correctness (PipeWire connection internals):
- LoopNode Q_DISABLE_COPY → Q_DISABLE_COPY_MOVE. Move would split the
  intrusive spa_hook from its listener list; std::unique_ptr-in-
  unordered_map ownership keeps the LoopNode pinned but a future
  refactor could trip move semantics silently.
- submitLoopRequest sync-dispatch double-free closed: release ownership
  to raw pointer BEFORE pw_loop_invoke, reclaim only on failure paths.
  Safe regardless of sync (loop-thread caller) vs async dispatch.
- bindAudioNode FIFO ordering: guiNodeAdded now queued BEFORE
  pw_node_enum_params + pw_node_subscribe_params. Closes the race
  where cached info from the proxy could synchronously trigger
  applyProps before the GUI node existed.
- onDefaultMetadataProperty parse-failure preservation: malformed JSON
  (non-object root, missing name field) now returns early without
  clobbering cached defaultSinkName/Name. Mirrors the oversize-payload
  hardening.
- isAudioNodeClass: replaced toUtf8 + memcmp with QStringView vs
  QLatin1String comparison. Allocation-free in the hot path.
- doParamWrite: qCDebug warning on volumes-array truncation.

P4 critical (PW models + qmlregistration):
- qmlregistration singleton factory: null-guard
  QCoreApplication::instance() (logs warning, returns nullptr) +
  hostPtr-vs-current-app validation so a dangling cache after app
  destruction gets re-created instead of being handed out.
- PipeWireHost: dropped dual ownership — connection is now solely owned
  by unique_ptr (no Qt parent). Removes the parent+unique_ptr
  destruction-order trap.
- PwNodeModel destructor: inline teardown without firing
  beginRemoveRows/endRemoveRows. Views handle model death via
  destroyed signal; signal emission during destruction risked
  partial-vtable callbacks.
- rebuildFromConnection: dropped newConnection parameter (Q_ASSERT
  enforced caller-must-match was debug-only; release builds could
  diverge). Uses d->connection member exclusively.
- Snapshot-seed loop: added rowIndex.contains guard for symmetry with
  the nodeAdded handler.
- PwNode::applyInfo: lazy device.description fallback (was always
  evaluated via function-arg eval).

Descoped polish (~33 MEDIUM + ~50 LOW, see audit reports):
- Header docstring drift (PipeWireHost forwarding gap, PwNode docstring
  setVolumes omission, etc.)
- Test robustness (objectName-based QTimer find vs findChildren,
  setDebounceIntervalForTest hook, daemon-side-affecting test opt-out
  env var)
- CLI ergonomics (range guard at usage layer vs runtime, install rule
  for shipping CLI as user-facing tool)
- Comment-code drift in palettestore.cpp file-top
- Defensive-dead-code patterns documented but not removed

These items are tracked in the audit reports captured across this and
prior invocations. They do not affect correctness or test coverage;
they are improvements that would emerge naturally from a doc-and-style
sweep PR rather than a correctness audit.
