+++ libs/phosphor-theme/tests/CMakeLists.txt
@@ -17,5 +17,12 @@ function(phosphortheme_add_test name)
 endfunction()
 
 phosphortheme_add_test(test_palettestore)
+# test_palettestore_files holds the loadFromFile slots; test_palettestore_hotreload
+# holds the hot-reload slots + the white-box resetToDefaults_releasesDirectoryWatch
+# slot. Split out of test_palettestore.cpp to keep every TU under the project's
+# 800-line cap. All three share tests/test_palettestore_helpers.h for the
+# QTemporaryDir + QFile fixture-write scaffolding.
+phosphortheme_add_test(test_palettestore_files)
+phosphortheme_add_test(test_palettestore_hotreload)
 phosphortheme_add_test(test_templateengine)
 phosphortheme_add_test(test_matugenrunner)
